requirements.txt
pytaglib==1.4.6
